requirements.txt
azure-storage-blob==1.3.1
